+++ CHANGELOG.md
@@ -68,6 +68,8 @@
 - Added the following CLI flags to control operator's behavior:
   - `--cache-sync-timeout` to control controller-runtime's time limit set to wait for syncing caches.
     [#1818](https://github.com/kong/kong-operator/pull/1818)
+  - `--cache-sync-period` to control controller-runtime's cache sync period.
+    [#1846](https://github.com/kong/kong-operator/pull/1846)
 
 ### Fixed